subscribie/requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Login==0.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gocardless-pro==1.5.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pyaml==17.12.1
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
